+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/vcs/githubcom/GithubComClient.kt
@@ -39,7 +39,6 @@ import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.withContext
 import org.slf4j.LoggerFactory
 import java.util.concurrent.atomic.AtomicInteger
-import java.util.regex.Pattern
 import kotlin.coroutines.CoroutineContext
 
 @SuppressWarnings("TooManyFunctions", "ReturnCount")
@@ -276,19 +275,11 @@ class GithubComClient @NonInjectable constructor(
 
     suspend fun validateAccess(searchHost: String, codeHost: String, settings: GitHubSettings): String = try {
         val client: HttpClient = httpClientProvider.getClient(searchHost, codeHost, settings)
-        val response: HttpResponse = client.get("${settings.baseUrl}/repos/mega-manipulator/mega-manipulator")
-        val scopeString = response.headers["X-OAuth-Scopes"]
-        val scopes = scopeString.orEmpty().split(Pattern.compile(",")).map { it.trim() }
-        println(scopes)
-        val expected = listOf("repo", "delete_repo")
-        val missing = expected - scopes
-        val missingText = if (missing.isNotEmpty()) "<br>\nmissing scopes: $missing" else ""
-        val rateLimit = response.headers["X-RateLimit-Limit"].orEmpty()
-        val rateLimitText: String = if (rateLimit == "") "<br>\nNo rate limit header in response, bad token?" else if ((rateLimit.toIntOrNull() ?: 0) < 100) "\nRateLimit is low, token is probably not setup correctly" else ""
-        "${response.status.value}:${response.status.description}$missingText$rateLimitText"
+        GitHubValidation.validateAccess(settings.baseUrl, client)
     } catch (e: Exception) {
-        logger.error("Failed request", e)
-        "Client error: ${e.message}"
+        val msg = "Failed setting up client: ${e.message}"
+        logger.error(msg, e)
+        msg
     }
 
     suspend fun approvePr(pullRequest: GithubComPullRequestWrapper, settings: GitHubSettings): PrActionStatus {